refactor: repository 명 수정
