+++ src/main/java/dynamicquad/agilehub/dummy/bulk/repository/ProjectBulkRepository.java
@@ -1,4 +1,4 @@
-package dynamicquad.agilehub.dummy.bulkRepository;
+package dynamicquad.agilehub.dummy.bulk.repository;
 
 import dynamicquad.agilehub.project.domain.Project;
 import java.sql.PreparedStatement;
